🏗️ Move doModProp into JsonSourceCopier
Let the base method handle the common mods, and then subclasses can
handle additional ones.
